# Use stage block txs table for ingesting raw transactions data into silver