Use the `bronze.stage_block_txs_2` to ingest data 
- Have the staging table hold up to ~7 days worth of data to make the process more resilient to data outages
- Update `silver.transactions` and `silver.votes` to pull from new upstream
- In some cases it will load a slightly different set of data due to how the raw file registration works but any "missed" blocks would get automatically re-requested. We can monitor this portion w/ the hourly tests and see if there is an increase in notifications there.


Iterating on same data as prod shows a significant performance boost in runtime

Transactions
```
Old
16:22:32  4 of 16 OK created sql incremental model silver.transactions ................... [SUCCESS 1030654 in 308.30s]
New
16:22:59  1 of 1 OK created sql incremental model silver.transactions .................... [SUCCESS 1030654 in 108.01s]

Old
16:38:04  25 of 171 OK created sql incremental model silver.transactions ................. [SUCCESS 1361976 in 330.59s]
New
16:35:25  1 of 1 OK created sql incremental model silver.transactions .................... [SUCCESS 1058891 in 92.16s]
```

Votes
```
Old
16:36:03  26 of 171 OK created sql incremental model silver.votes ........................ [SUCCESS 6164789 in 209.52s]
New
16:33:18  1 of 1 OK created sql incremental model silver.votes ........................... [SUCCESS 5326780 in 53.58s]
```